third_party/WebKit/LayoutTests/fast/forms/select/menulist-change-event-with-reset-blur.html

<!DOCTYPE html>
<body>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="log"></div>
<form>
<select>
<option selected>1</option>
<option>2</option>
</select>
</form>
<script>
var select = document.querySelector('select');
var changeEventCounter = 0;
select.addEventListener('change', function() { ++changeEventCounter; }, false);

test(function() {
    select.focus();
    eventSender.keyDown('2');
    assert_equals(changeEventCounter, 1);
    document.querySelector('form').reset();
    select.blur();
    assert_equals(changeEventCounter, 1);
}, 'Form reset and blur() should trigger no change event.');
</script>
</body>
